Move coverage tests into their respective per-class test files
The combined CoverageGapTests.cs file deviated from the repo convention of
one test class per production class. Tests have been merged into the
appropriate existing test files (CardinalityEstimatorTests,
CardinalityEstimatorSerializerTests, ConcurrentCardinalityEstimatorTests,
CardinalityEstimatorExtensionsTests, Fnv1ATests, Murmur3Tests).

No behavior change; full suite still passes 177/1.
